+++ src/nespresso/api/routers/nes_user.py
@@ -7,7 +7,7 @@ from fastapi.security import APIKeyHeader
 from nespresso.core.configs.settings import settings
 from nespresso.db.models.nes_user import NesUser
 from nespresso.db.schemas.nes_user import NesUserIn, NesUserOut
-from nespresso.db.services.user_context import user_ctx
+from nespresso.db.services.user_context import GetUserContextService
 
 router = APIRouter()
 
@@ -52,7 +52,7 @@ async def UpsertNesUser(
     ],
     token: Annotated[str, Depends(VerifyMyNesToken)],
 ) -> NesUserOut:
-    ctx = await user_ctx()
+    ctx = await GetUserContextService()
 
     alchemy_nes_user: NesUser = NesUserPydanticToSQLAlchemy(nes_user)