+++ src/nespresso/bot/lib/message/checks.py
@@ -1,9 +1,9 @@
 from nespresso.db.models.tg_user import TgUser
-from nespresso.db.services.user_context import user_ctx
+from nespresso.db.services.user_context import GetUserContextService
 
 
 async def CheckVerified(chat_id: int) -> bool:
-    ctx = await user_ctx()
+    ctx = await GetUserContextService()
 
     verified = await ctx.GetTgUser(
         chat_id=chat_id,