+++ src/nespresso/core/logs/api.py
@@ -11,11 +11,6 @@ from nespresso.core.logs.settings import (
 
 
 async def LoggerSetup() -> QueueListener:
-    logging.getLogger("sqlalchemy.engine").setLevel(logging.WARNING)
-    logging.getLogger("opensearch").setLevel(logging.INFO)
-    logging.getLogger("filelock").setLevel(logging.INFO)
-
-    # ─── API Handlers ───
     console_handler = CreateConsoleHandler(
         logging.INFO,
         filters=[
@@ -27,7 +22,6 @@ async def LoggerSetup() -> QueueListener:
         logging.DEBUG,
     )
 
-    # ─── Queue ───
     listener: QueueListener = CreateListener(console_handler, bot_file_handler)
 
     return listener